requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
configparser==3.5.0
cryptography==2.3.1
enum34==1.1.6
idna==2.7
ipaddress==1.0.22
logging==0.4.9.6
pycparser==2.18
PyMySQL==0.9.2
PyYAML==3.13
six==1.11.0
virtualenv==16.0.0
